+++ Source/Core Providers/Core_Midi/Core_Midi.cs
@@ -21,7 +21,7 @@ namespace Core_Midi
             = new ConcurrentDictionary<DeviceDescriptor, IDeviceHandler<MidiInMessageEventArgs>>();
         private readonly ConcurrentDictionary<DeviceDescriptor, MidiOutputDeviceHandler> _activeOutputDevices
             = new ConcurrentDictionary<DeviceDescriptor, MidiOutputDeviceHandler>();
-        private Action<ProviderDescriptor, DeviceDescriptor, BindingReport, int> _bindModeCallback;
+        private Action<ProviderDescriptor, DeviceDescriptor, BindingReport, short> _bindModeCallback;
         private readonly object _lockObj = new object();  // When changing mode (Bind / Sub) or adding / removing devices, lock this object
 
         public Core_Midi()
@@ -54,7 +54,7 @@ namespace Core_Midi
             _deviceLibrary.RefreshConnectedDevices();
         }
 
-        public void SetDetectionMode(DetectionMode detectionMode, DeviceDescriptor deviceDescriptor, Action<ProviderDescriptor, DeviceDescriptor, BindingReport, int> callback = null)
+        public void SetDetectionMode(DetectionMode detectionMode, DeviceDescriptor deviceDescriptor, Action<ProviderDescriptor, DeviceDescriptor, BindingReport, short> callback = null)
         {
             lock (_lockObj)
             {
