+++ Source/Core Providers/Core_SpaceMouse/Core_SpaceMouse.cs
@@ -21,7 +21,7 @@ namespace Core_SpaceMouse
         private readonly ConcurrentDictionary<DeviceDescriptor, IDeviceHandler<HidReport>> _activeDevices
             = new ConcurrentDictionary<DeviceDescriptor, IDeviceHandler<HidReport>>();
         private readonly IInputDeviceLibrary<string> _deviceLibrary;
-        private Action<ProviderDescriptor, DeviceDescriptor, BindingReport, int> _bindModeCallback;
+        private Action<ProviderDescriptor, DeviceDescriptor, BindingReport, short> _bindModeCallback;
         private readonly ProviderDescriptor _providerDescriptor;
         private readonly object _lockObj = new object();  // When changing mode (Bind / Sub) or adding / removing devices, lock this object
 
@@ -51,7 +51,7 @@ namespace Core_SpaceMouse
             _deviceLibrary.RefreshConnectedDevices();
         }
 
-        public void SetDetectionMode(DetectionMode detectionMode, DeviceDescriptor deviceDescriptor, Action<ProviderDescriptor, DeviceDescriptor, BindingReport, int> callback = null)
+        public void SetDetectionMode(DetectionMode detectionMode, DeviceDescriptor deviceDescriptor, Action<ProviderDescriptor, DeviceDescriptor, BindingReport, short> callback = null)
         {
             lock (_lockObj)
             {
